requirements.txt
torch
numpy
argparse
pyyaml
absl-py==1.4
gin-config==0.5
open3d==0.17
pandas==2.0
scipy==1.10
tensorflow==2.13
tensorflow-datasets==4.9
tensorflow-graphics==2021.12.2
tensorflow-probability==0.21
tf-models-official==2.5
tqdm==4.66
